feat: Implement JSON serialization for runtime persistence layer
- Added `runtime_serde.rs` to handle JSON conversion for `DataValue`, `Content`, and `TaskResult`.
- Introduced functions for converting tool step inputs and outputs to JSON format.
- Enhanced task result persistence with JSON serialization and deserialization.
- Updated memory management to use `u64` for elapsed time instead of `u128`.
- Refactored timestamp handling across various structs to use `u64`.
- Improved error handling in round-robin and supervisor strategies to capture and report turn errors.
- Added elapsed time tracking for task execution in strategies.

+++ hosts/rust-daemon/src/routes/contracts/memories.rs
@@ -29,7 +29,7 @@ pub(crate) struct MemoryResponse {
     pub(crate) memory_type: String,
     pub(crate) content: String,
     pub(crate) importance: f64,
-    pub(crate) created_at: u128,
+    pub(crate) created_at: u64,
     pub(crate) tags: Option<Vec<String>>,
     pub(crate) scope: String,
     pub(crate) room_id: Option<String>,
@@ -47,7 +47,7 @@ pub(crate) struct MemorySearchResultResponse {
     pub(crate) memory_type: String,
     pub(crate) content: String,
     pub(crate) importance: f64,
-    pub(crate) created_at: u128,
+    pub(crate) created_at: u64,
     pub(crate) tags: Option<Vec<String>>,
     pub(crate) score: f64,
     pub(crate) scope: String,
@@ -85,8 +85,8 @@ pub(crate) struct AgentRelationshipResponse {
     pub(crate) room_id: Option<String>,
     pub(crate) world_id: Option<String>,
     pub(crate) session_id: Option<String>,
-    pub(crate) created_at: u128,
-    pub(crate) updated_at: u128,
+    pub(crate) created_at: u64,
+    pub(crate) updated_at: u64,
 }
 
 #[derive(Clone, Debug, Serialize, ToSchema)]
@@ -102,8 +102,8 @@ pub(crate) struct MemoryEntityResponse {
     pub(crate) name: String,
     pub(crate) aliases: Vec<String>,
     pub(crate) summary: Option<String>,
-    pub(crate) created_at: u128,
-    pub(crate) updated_at: u128,
+    pub(crate) created_at: u64,
+    pub(crate) updated_at: u64,
 }
 
 #[derive(Clone, Debug, Serialize, ToSchema)]
@@ -282,10 +282,10 @@ pub(crate) struct MemoryEvaluationRequest {
 #[derive(Clone, Debug, Serialize, Deserialize, ToSchema)]
 #[serde(rename_all = "camelCase")]
 pub(crate) struct MemoryRetentionRequest {
-    pub(crate) max_age_millis: Option<u128>,
+    pub(crate) max_age_millis: Option<u64>,
     pub(crate) min_importance: Option<f64>,
     pub(crate) max_memories: Option<usize>,
-    pub(crate) decay_half_life_millis: Option<u128>,
+    pub(crate) decay_half_life_millis: Option<u64>,
 }
 
 #[derive(Clone, Debug, Deserialize, IntoParams, ToSchema, Default)]
